LICENSE.md: update license years and vendor text
We last updated our software license in commit
4da6d5f30c52811411b9db796442ec1f72ae8a1d of PR #4513 to revise the
current year in our copyright statement to 2021.

As this date is now several years in the past, and because we want to
avoid having to update our license each year, we simply remove the
current year from our license's copyright statement now.

This practice is in keeping with the format we use in the license files
of all our other Git LFS projects, such as in the git-lfs/gitobj and
git-lfs/pktline repositories.  It also aligns with the practice of other
major open-source projects, like that of the source code for the Go
language and standard library.

As well, we add a copyright date of 2009 to the Go project's license
that we include in our main LICENSE.md, alongside the existing 2010
copyright date.  When we first appended the Go project's license to our
license file, in commit 74d5f2397f9abe4834bf1fe1fa02fd6c141b77cea, it was
specifically due to the inclusion of some code into our "subprocess"
package from the Go library's "os/exec" package.  Those original source
files have copyright dates of 2010, which is why we used that date in
our copy of the Go project's license.  For reference, see:

  https://github.com/golang/go/blob/b68f8ca89a2fb6b33d1e78fadc33d1c35693f6b8/src/os/exec/lp_unix.go#L1
  https://github.com/golang/go/blob/b68f8ca89a2fb6b33d1e78fadc33d1c35693f6b8/src/os/exec/lp_windows.go#L1

In commit a66274014c5c1d8dcc0d9ccbca61bf75eb792995 of PR #4418 we then
revised our license file to note that we also include some code from the
Go library's "os" package in our "tools" package.  The Go source code
in this case has a copyright date of 2009, but we did not update our
license file to reflect this, so we do that now.  For reference, see:

  https://github.com/golang/go/blob/b68f8ca89a2fb6b33d1e78fadc33d1c35693f6b8/src/os/file_windows.go#L1

Finally, in commit fc95175ce773941516c8fe600549aa37f53048d6 of PR #4903
we removed our "vendor" directory and all the associated copies of our
dependencies from this project.  Our license file still mentions this
directory, though, so we remove that reference now.