creds/creds.go: reject nulls in credential data
The protocol for communication with the git-credential(1) command,
as specified in the Git documentation, permits all bytes except for
null (NUL) and "newline" control characters, where "newline" is intended
to mean a line feed (LF) character that may possibly be preceded by a
carriage return (CR) character:

  https://github.com/git/git/blob/v2.47.1/Documentation/git-credential.txt#L116-L117

When a credential request is made to the "git credential fill" command
with a line feed character in one of the input values, Git will detect
the character and return an error.  This check is implemented in the
credential_write_item() function of the credential.c source file:

  https://github.com/git/git/blob/v2.47.1/credential.c#L391-L392

(In addition, as described in prior commits in this PR, future versions
of Git will also disallow bare carriage return (CR) characters, unless
the new "credential.protectProtocol" configuration option is set to a
value equivalent to "false".)

However, no such check is performed for null bytes.  If an input line
contains a null byte then Git will silently discard that byte and any
data following it, through to the end of the input line, because the
credential_read() function casts the input line's buffer as a null-
terminated C string:

  https://github.com/git/git/blob/v2.47.1/credential.c#L312-L313

The consequence is that the git-credential(1) command simply ignores
null bytes.  Nevertheless, to be compliant with the protocol's
specification we now adjust Git LFS so as to ensure that we never
pass any null bytes in the input we send to the command.

Following the pattern established by our previous commits in this PR,
we add an additional check in the buffer() method of the Creds structure
of our "creds" package.  This check returns an error if a null byte is
found in any of the values the method would otherwise write to its buffer.

We then update the corresponding TestCredsBufferProtect() test function
to confirm that an error is returned by the buffer() method when a null
byte is included in one of the values passed to it, regardless of whether
the method's protectProtocol argument is set to "false" or not.

Finally, we add a "credentials rejected with null byte" test to the
t/t-credentials-protect.sh test script.  This new test is nearly
identical to the "credentials rejected with line feed" test we
introduced in a prior commit in this PR.  The tests only differ in the
specific error messages they expect and, of course, in the use of a
URL-encoded null byte in our new test's Git LFS configuration URL
instead of a URL-encoded line feed character.