script/build-git: only link curl and zlib on macOS
Since we migrated our CI jobs to GitHub Actions in PR #3808, our
"build-latest" and "build-earliest" jobs, when running on macOS, have
explicitly installed and linked the "curl", "openssl", "pcre2", and
"zlib" Homebrew formulae before building a custom installation of Git.

At present, the macOS runners provided by GitHub Actions are always
provisioned with pre-installed versions of recent releases of the curl,
OpenSSL, and PCRE2 libraries, as packaged by Homebrew, so we no longer
need to install the Homebrew formulae for them.

We also do not need to link the "openssl" and "pcre2" formulae, but
because macOS includes its own default installation of curl and its
library, we continue to run "brew link" for the "curl" formula (as
well as "zlib") so that Homebrew's version of curl is used in
preference to the one installed by macOS.