.github/workflows: skip gettext install on macOS
Since commit 628571695fde7ec92b6d18ae87d8d7e4dc87d5d4 in PR #4729 we
have installed the gettext library and utilities in our CI and release
GitHub Actions jobs, on both macOS and Ubuntu Linux platforms, while
on Windows we rely on the library and utilities being installed in
advance as part of the Git for Windows environment.

At present, though, the macOS runners provided by GitHub Actions are
always provisioned with a pre-installed version of a recent release of
the gettext library, as packaged by Homebrew, which includes the
msgfmt(1) program used by our Makefile, so we no longer need to install
the "gettext" Homebrew formula ourselves.