go.*: require full Go version for toolchain
Since Go version 1.21, Go supports the designation of a specific version
of the Go toolchain to use when building a module, as described in:

  https://go.dev/doc/toolchain

If the "go" line in a "go.mod" file contains a full version, i.e., one
in the M.N.P format, then this will suffice to define the minimum
required Go toolchain to use when building the module.  Otherwise, Go
will look for a "toolchain" line in the "go.mod" file.

As we do not yet provide either a "toolchain" line or a "go" line with
the full version, running "go get" and other Go utilities may attempt
to insert a "toolchain" line with a recent Go version, such as 1.23.3.
We also receive warnings from CodeQL that our "go.mod" file contains
an invalid toolchain version.

To avoid these issues, we specify a full Go version in the "go" line
of our "go.mod" file, which suffices to define the minimum toolchain
version as well.

We then also run "go mod tidy" to update our module requirements,
which removes a number of entries that have become unnecessary.