workflows: improve shell quoting
As it stands, the architecture and the container are not
user-controllable and do not presently contain spaces or other shell
metacharacters, so there's no security or functionality problems by
leaving the relevant environment variables unquoted.  However, it's a
good practice to quote these variables just in case, so let's do that.