workflows: don't persist credentials
By default, the `actions/checkout` action stores the credentials in the
local Git config file.  This is not really very secure and a credential
manager would be a much better idea, but Actions has not yet done so.

In the mean time, since we don't need those credentials to perform more
Git operations, let's make sure to not persist them in the config,
which means that it's less likely we'll accidentally expose them, such
as by shipping them in an artifact.