script/build-git: log Git build dependencies
As suggested by larsxschneider on PR review, to make it easier to
diagnose problems with our "build-earliest" and "build-latest" CI jobs
we can log the build options and dynamic library dependencies of the
custom versions of Git that we compile in those jobs.

The --build-options argument to the git-version(1) command was introduced
with Git v2.46.0, but will be silently ignored by earlier versions, so
we can use it directly in the script/build-git script called by our
"build-earliest" and "build-latest" CI jobs, including the former, which
builds Git version 2.0.0.

In addition, we can also list the dynamic library dependencies of the
"git-http-fetch" binary we build, since this is one of the Git programs
which must be linked with the libcurl library and we depend on this Git
program in our test suite.  We therefore use either the ldd(1) command
on Linux or the otool(1) command on macOS to display the dependencies
of the "git-http-fetch" binary after we have built and installed it.