Add `--dry-run` option to `fetch`
This was modelled after the `git-lfs push --dry-run` option. I've
removed some apparently unused code (computing and reporting of ready
nodes, and the output channel of the main fetch function which was
always `nil`).

For testing, I added `--dry-run` testing to a couple of cases, but I
wonder if we should add it for _all_ cases.

This is a part of what was discussed in
https://github.com/git-lfs/git-lfs/issues/5888.