.github/workflows,docs/howto: manual asset uploads
Since PR #3840 introduced a GitHub Actions workflow to automate much
of the process of releasing a new version of Git LFS, that workflow
has concluded by running the script/upload script, which makes requests
to GitHub's API to create a new draft release announcement and attach
the binary release assets built by the workflow to the announcement.

This step necessarily requires that the GITHUB_TOKEN credential provided
to the workflow have permission to create a release, which implies
fairly broad write access to the repository contents as GitHub does
not offer a more fine-grained permission model for this particular task.

In order to ensure that our GitHub Actions workflows can not be
misused, we would much prefer to run them without write permissions to
our repositories.  We therefore revise the "build-main" job in our
release workflow to remove the step which tries to run the script/upload
script, and replace it with steps that instead create a "release-assets"
artifact containing the binary release assets built by the workflow.

The administrator performing the release may then download this artifact
and unpack it, after which they can run the script/upload script manually
using their own personal GitHub credentials.  We update our release
process documentation to explain this new approach and also clarify a
few other details of the release process.

This change does make our release process slightly more manual, and
if in the future GitHub provides a convenient way to enable a single
job to have just sufficient permissions to create a draft release
announcement, we can revisit this decision in the future.