Makefile: use macOS keychain ID consistently
In commit bfc5304edf2b218fb8abf22aba02b5dc7e842a4a of PR #4143 we
introduced a set of Makefile targets which sign and notarize binaries
on macOS using the "codesign" and "security" utilities, among others.
The recipes for these targets create a new, dedicated keychain and
add our Apple Developer ID certificate to it, and then use the keychain
to sign the macOS binaries we build.

Although we refer to this keychain in some commands using the Makefile
variable DARWIN_KEYCHAIN_ID, whose default is the value "CI.keychain",
in other commands we refer to that keychain name explicitly.  As a
consequence, the variable can not actually be set to something other
than its default value without causing problems.

As we expect to further revise some of these Makefile targets in
subsequent commits in this PR, we first adjust their recipes to
use the DARWIN_KEYCHAIN_ID variable consistently, and we also take
the opportunity to alter its default value to "lfs.keychain", since
we run these recipes in both our CI and release GitHub Actions
workflows, not just in the CI workflow.