Makefile: simplify macOS signing log output
In commit bfc5304edf2b218fb8abf22aba02b5dc7e842a4a of PR #4143 we
introduced a set of Makefile targets which sign and notarize binaries
on macOS using the "codesign" and "security" utilities, among other
tools.  The recipes for these targets create a new, dedicated keychain
and add our Apple Developer ID certificate to it, and then use the
keychain to sign the macOS binaries we build.

These recipes include several commands which generate a considerable
amount of detailed diagnostic log output, making the logs somewhat
difficult to read.  We can clarify the steps involved in the signing
process by adding several log statements, removing commands which
only generate extra diagnostics, and silencing the output of others.

If we encounter problems with the macOS signing process in the
future, we can always reinstate these commands while running the
Makefile recipes on a local machine or in a private repository
in order to trace the source of the issues.