.github/workflows,Makefile: find macOS cert ID
In commit bfc5304edf2b218fb8abf22aba02b5dc7e842a4a of PR #4143 we
introduced a set of Makefile targets which sign and notarize binaries
on macOS using the "codesign" and "security" utilities, among other
tools.  The recipes for these targets create a new, dedicated keychain
and add our Apple Developer ID certificate to it, and then use the
keychain to sign the macOS binaries we build.

At present, we retrieve our Apple Developer ID certificate and password
from GitHub Actions secrets, along with our Developer ID signing identity,
which is provided in an Actions secret named MACOS_CERT_ID.  We then pass
that identity to the "codesign" command with the -s option (i.e., the
--sign option) in order to sign our binary release assets.

In a subsequent commit in this PR we will modify our release GitHub
Actions workflow to retrieve the certificate data from a new system,
as our current Apple Developer ID certificate expired recently and
can not simply be renewed.  This new system to manage Apple Developer
ID certificates does not provide the signing identity in an Actions
secret, just the certificate and password.

Fortunately, we can retrieve the signing identity from the local macOS
keychain we have created in the recipe for our "release-import-certificate"
Makefile target, using the "find-identity" subcommand of the "security"
command, so we do that now at the start of our "release-darwin" target's
recipe.

This in turn allows us to remove the DARWIN_CERT_ID variable from our
Makefile and update our release workflow as it no longer needs to set
that environment variable from the MACOS_CERT_ID Actions secret.