.github/workflows,docs/howto: update macOS signing
In commit bfc5304edf2b218fb8abf22aba02b5dc7e842a4a of PR #4143 we
introduced a set of Makefile targets which sign and notarize binaries
on macOS using the "codesign" and "security" utilities, among other
tools.  The recipes for these targets create a new, dedicated keychain
and add our Apple Developer ID certificate to it, and then use the
keychain to sign the macOS binaries we build.

Our most recent certificate expired recently, and per the current
Apple documentation, a maximum of five certificates are permitted
for each Apple Developer ID:

  You can create up to five Developer ID Application certificates and
  up to five Developer ID Installer certificates using either your
  developer account or Xcode.

  https://developer.apple.com/help/account/create-certificates/create-developer-id-certificates

As our Developer ID is provided to us by GitHub, and they distribute
multiple other projects which also need to be signed, we are no longer
able to use a certificate dedicated just to Git LFS, and so need to
adopt GitHub's process for managing these certificates.  To do so we
alter our release GitHub Actions workflow to retrieve the necessary
certificate and password from the Actions secrets provided by GitHub's
internal tooling.  This tool makes the certificate data available in
Actions environment secrets using an environment named "production",
so we add that environment to most of the jobs in our release workflow.

When we push a new release tag and our release workflow is enqueued,
it will now wait to begin running these jobs until we manually validate
the workflow in GitHub's Apple Developer ID certificate management
interface, so we add that step to our release process documentation.

Note that we exclude the "build-main" job in our release workflow
from the "production" environment because it runs only after the
"build-macos" and "build-windows" jobs are complete, and if it ran
in the "production" environment it would therefore require a second
manual validation in the GitHub certificate management tool.  As this
would be inconvenient and unnecessary, we just allow the job to run
in the default environment.